requirements.txt
flask
flask-sqlalchemy
flask-mail
flask-cors
python-dotenv
gunicorn
psycopg2-binary



